+++ superduper/rest/build.py
@@ -280,10 +280,9 @@ def build_rest_app(app: SuperDuperApp):
     @app.add('/db/show_template', method='get')
     def db_show_template(
         identifier: str,
-        type_id: str = 'template',
         db: 'Datalayer' = DatalayerDependency(),
     ):
-        template: Template = db.load(type_id=type_id, identifier=identifier)
+        template: Template = db.load(type_id='template', identifier=identifier)
         return template.form_template
 
     @app.add('/db/edit', method='get')
@@ -293,7 +292,7 @@ def build_rest_app(app: SuperDuperApp):
         db: 'Datalayer' = DatalayerDependency(),
     ):
         component = db.load(type_id, identifier)
-        template = db.load('template', component.build_template)
+        template: Template = db.load('template', component.build_template)
         form = template.form_template
         form['_variables'] = component.build_variables
         return form